Rename CollectionOfUcustomers[1].java to Utility Customer List.java
This contains a client class, called CollectionOfUcustomers, that maintains a list of UtilityCustomer objects created by a user. The client class prompts a user for the type of a UtilityCustomer and its corresponding parameters
to create (ex. bill, type of customer, etc). The program allows the user to specify up to 10 UtilityCustomer objects to be stored in an array. The program displays the information of UtilityCustomer objects in descending order based on their account numbers.